@@ -1,5 +1,5 @@
 from functools import partial
-from typing import Any, ClassVar, Dict, Literal, Optional, Tuple, Type, Union
+from typing import Any, ClassVar, Literal, Optional, Union
 
 import flax
 import flax.linen as nn
@@ -16,7 +16,7 @@ from stable_baselines3.common.type_aliases import GymEnv, MaybeCallback, Schedul
 
 from sbx.common.off_policy_algorithm import OffPolicyAlgorithmJax
 from sbx.common.type_aliases import ReplayBufferSamplesNp, RLTrainState
-from sbx.sac.policies import SACPolicy
+from sbx.sac.policies import SACPolicy, SimbaSACPolicy
 
 
 class EntropyCoef(nn.Module):
@@ -40,8 +40,10 @@ class ConstantEntropyCoef(nn.Module):
 
 
 class SAC(OffPolicyAlgorithmJax):
-    policy_aliases: ClassVar[Dict[str, Type[SACPolicy]]] = {  # type: ignore[assignment]
+    policy_aliases: ClassVar[dict[str, type[SACPolicy]]] = {  # type: ignore[assignment]
         "MlpPolicy": SACPolicy,
+        # Residual net, from https://github.com/SonyResearch/simba
+        "SimbaPolicy": SimbaSACPolicy,
         # Minimal dict support using flatten()
         "MultiInputPolicy": SACPolicy,
     }
@@ -60,19 +62,21 @@ class SAC(OffPolicyAlgorithmJax):
         batch_size: int = 256,
         tau: float = 0.005,
         gamma: float = 0.99,
-        train_freq: Union[int, Tuple[int, str]] = 1,
+        train_freq: Union[int, tuple[int, str]] = 1,
         gradient_steps: int = 1,
         policy_delay: int = 1,
         action_noise: Optional[ActionNoise] = None,
-        replay_buffer_class: Optional[Type[ReplayBuffer]] = None,
-        replay_buffer_kwargs: Optional[Dict[str, Any]] = None,
+        replay_buffer_class: Optional[type[ReplayBuffer]] = None,
+        replay_buffer_kwargs: Optional[dict[str, Any]] = None,
         ent_coef: Union[str, float] = "auto",
         target_entropy: Union[Literal["auto"], float] = "auto",
         use_sde: bool = False,
         sde_sample_freq: int = -1,
         use_sde_at_warmup: bool = False,
+        stats_window_size: int = 100,
         tensorboard_log: Optional[str] = None,
-        policy_kwargs: Optional[Dict[str, Any]] = None,
+        policy_kwargs: Optional[dict[str, Any]] = None,
+        param_resets: Optional[list[int]] = None,  # List of timesteps after which to reset the params
         verbose: int = 0,
         seed: Optional[int] = None,
         device: str = "auto",
@@ -96,7 +100,9 @@ class SAC(OffPolicyAlgorithmJax):
             use_sde=use_sde,
             sde_sample_freq=sde_sample_freq,
             use_sde_at_warmup=use_sde_at_warmup,
+            stats_window_size=stats_window_size,
             policy_kwargs=policy_kwargs,
+            param_resets=param_resets,
             tensorboard_log=tensorboard_log,
             verbose=verbose,
             seed=seed,
@@ -189,6 +195,9 @@ class SAC(OffPolicyAlgorithmJax):
         # Sample all at once for efficiency (so we can jit the for loop)
         data = self.replay_buffer.sample(batch_size * gradient_steps, env=self._vec_normalize_env)
 
+        # Maybe reset the parameters/optimizers fully
+        self._maybe_reset_params()
+
         if isinstance(data.observations, dict):
             keys = list(self.observation_space.keys())  # type: ignore[attr-defined]
             obs = np.concatenate([data.observations[key].numpy() for key in keys], axis=1)
@@ -211,7 +220,7 @@ class SAC(OffPolicyAlgorithmJax):
             self.policy.actor_state,
             self.ent_coef_state,
             self.key,
-            (actor_loss_value, qf_loss_value, ent_coef_value),
+            (actor_loss_value, qf_loss_value, ent_coef_loss_value, ent_coef_value),
         ) = self._train(
             self.gamma,
             self.tau,
@@ -229,6 +238,7 @@ class SAC(OffPolicyAlgorithmJax):
         self.logger.record("train/n_updates", self._n_updates, exclude="tensorboard")
         self.logger.record("train/actor_loss", actor_loss_value.item())
         self.logger.record("train/critic_loss", qf_loss_value.item())
+        self.logger.record("train/ent_coef_loss", ent_coef_loss_value.item())
         self.logger.record("train/ent_coef", ent_coef_value.item())
 
     @staticmethod
@@ -291,7 +301,7 @@ class SAC(OffPolicyAlgorithmJax):
     ):
         key, dropout_key, noise_key = jax.random.split(key, 3)
 
-        def actor_loss(params: flax.core.FrozenDict) -> Tuple[jax.Array, jax.Array]:
+        def actor_loss(params: flax.core.FrozenDict) -> tuple[jax.Array, jax.Array]:
             dist = actor_state.apply_fn(params, observations)
             actor_actions = dist.sample(seed=noise_key)
             log_prob = dist.log_prob(actor_actions).reshape(-1, 1)
@@ -382,10 +392,11 @@ class SAC(OffPolicyAlgorithmJax):
                 "actor_loss": jnp.array(0.0),
                 "qf_loss": jnp.array(0.0),
                 "ent_coef_loss": jnp.array(0.0),
+                "ent_coef_value": jnp.array(0.0),
             },
         }
 
-        def one_update(i: int, carry: Dict[str, Any]) -> Dict[str, Any]:
+        def one_update(i: int, carry: dict[str, Any]) -> dict[str, Any]:
             # Note: this method must be defined inline because
             # `fori_loop` expect a signature fn(index, carry) -> carry
             actor_state = carry["actor_state"]
@@ -429,7 +440,12 @@ class SAC(OffPolicyAlgorithmJax):
                 target_entropy,
                 key,
             )
-            info = {"actor_loss": actor_loss_value, "qf_loss": qf_loss_value, "ent_coef_loss": ent_coef_loss_value}
+            info = {
+                "actor_loss": actor_loss_value,
+                "qf_loss": qf_loss_value,
+                "ent_coef_loss": ent_coef_loss_value,
+                "ent_coef_value": ent_coef_value,
+            }
 
             return {
                 "actor_state": actor_state,
@@ -446,5 +462,10 @@ class SAC(OffPolicyAlgorithmJax):
             update_carry["actor_state"],
             update_carry["ent_coef_state"],
             update_carry["key"],
-            (update_carry["info"]["actor_loss"], update_carry["info"]["qf_loss"], update_carry["info"]["ent_coef_loss"]),
+            (
+                update_carry["info"]["actor_loss"],
+                update_carry["info"]["qf_loss"],
+                update_carry["info"]["ent_coef_loss"],
+                update_carry["info"]["ent_coef_value"],
+            ),
         )
